requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==4.1.4
django-environ==0.7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
django-paranoid-model==1.0.7
drf-yasg==1.20.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==22.0
psycopg2==2.8.6
PyJWT==2.1.0
pytz==2022.6
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.13
